Added frontend files
In this upload, I have a basic but not yet finished front end of my artist portfolio.  
Features: 
Ability to navigate through multi-view applications. 
Subpages are stylized with a grid layout 
Grid items can be clicked on to lead to another subpage of the expanded image. 
About me, section is finished. 
Ideas form is created, however data collection is not yet implemented. 
Local Storage is utilized in subpage and artwork views.